+++ SoD/Level-40-Horde.lua
@@ -1,0 +1,57 @@
+RXPGuides.RegisterGuide([[
+#classic
+#version 1
+<< Horde
+#group Custom Quests for RXP
+#subgroup SoD Level 40 Quest Stacking
+#name 40-1-How to Use this Guide
+#displayname 1-How to Use this Guide
+#next 40-2-Pre-Launch Collection Guide
+
+-- |cRXP_FRIENDLY_, |cRXP_ENEMY_, |cRXP_LOOT_, |cRXP_WARN_, |cRXP_PICK_, |cRXP_BUY_
+-- |c99ffff99 OR|r
+-- |Tinterface/worldmap/chatbubble_64grey.blp:20|tTalk
+
+step
+    >>Welcome to the |cRXP_WARN_SoD Level 40|r Quest Stacking Guide!
+    +Use the Collection Guide to complete pre-req quests and gather items for launch day.
+    +|cRXP_WARN_ON LAUNCH DAY|r, use the Turn-in Guide.
+]])
+
+RXPGuides.RegisterGuide([[
+#classic
+#version 1
+<< Horde
+#group Custom Quests for RXP
+#subgroup SoD Level 40 Quest Stacking
+#name 40-2-Pre-Launch Collection Guide
+#displayname 2-Pre-Launch Collection Guide
+#next 40-3-Launch Day Turn-in Route
+
+step
+    .maxlevel 39
+    >>|cRXP_WARN_NOTE|r: This guide is for |cRXP_WARN_HIGHER LEVEL|r characters.
+    .xp 40
+step
+    >>|cRXP_WARN_LEVEL 40 HORDE QUEST STACK|r
+    >>There is |cRXP_PICK_NO RXP|r Horde collection guide for Phase 3.
+    +Use the Google Sheet and Weak Aura available through:
+    .link https://www.patreon.com/joardee >>Joardee's Patreon
+]])
+
+RXPGuides.RegisterGuide([[
+#classic
+#version 1
+<< Horde
+#group Custom Quests for RXP
+#subgroup SoD Level 40 Quest Stacking
+#name 40-3-Launch Day Turn-in Route
+#displayname 3-Launch Day Turn-in Route
+
+step
+    +|cRXP_WARN_WARNING: ONLY LOAD THIS GUIDE ON LAUNCH DAY, DO NOT ACCIDENTALLY TURN IN YOUR PRE QUESTS|r
+step
+    >>|cRXP_WARN_LEVEL 40 HORDE QUEST STACK|r
+    >>There is no guarantee an RXP turn-in guide will be available before Phase 3 launch.
+    +Check back later for updates
+]])